Q_Evil Smile_Q
QQ.py is meant for breaking down into constrained sub problems a genetic chain allowing the use of the thoroughest GA heuristic one can imagine.  An illustration is made for a splitting instance of random letters with constrains. Then an induction is possible for larger problems that comprise an optimization scenario. It was tested first, in a data structure that required only 2 parameters to be fully expounded. A shipyard monitoring system that is in which it worked flawllesly and under any case. QRNL uses Instr4,5 while QQ uses Instr1,2,3, respectively. The prototype stemed from QRNL while test was created from the GA split of the QQ, the only real challenge was to impose explicit constrains when moving from the double description to a triple description. A triple description allows the user to include instructions that would be otherwise trivias.